Remove amp-video-service experiment
