--- extensions/amp-video-service/OWNERS.yaml
@@ -1,1 +1,0 @@
-- alanorozco
