+++ src/services.js
@@ -571,12 +571,11 @@ export class Services {
 
   /**
    * @param {!Element|!./service/ampdoc-impl.AmpDoc} elementOrAmpDoc
-   * @return {!./service/video-service-interface.VideoServiceInterface}
+   * @return {!./service/video-manager-impl.VideoManager}
    */
   static videoManagerForDoc(elementOrAmpDoc) {
-    return (
-      /** @type {!./service/video-service-interface.VideoServiceInterface} */ (
-        getServiceForDoc(elementOrAmpDoc, 'video-manager')));
+    return (/** @type {!./service/video-manager-impl.VideoManager} */ (
+      getServiceForDoc(elementOrAmpDoc, 'video-manager')));
   }
 
   /**
