+++ src/video-interface.js
@@ -491,3 +491,22 @@ export let VideoOrBaseElementDef;
 export function isDockable(element) {
   return element.hasAttribute(VideoAttributes.DOCK);
 }
+
+
+/** @enum {string} */
+export const VideoServiceSignals = {
+  USER_INTERACTED: 'user-interacted',
+  AUTOPLAY_DELEGATED: 'autoplay-delegated',
+};
+
+
+/** @param {!AmpElement|!VideoOrBaseElementDef} video */
+export function delegateAutoplay(video) {
+  video.signals().signal(VideoServiceSignals.AUTOPLAY_DELEGATED);
+}
+
+
+/** @param {!AmpElement|!VideoOrBaseElementDef} video */
+export function userInteractedWith(video) {
+  video.signals().signal(VideoServiceSignals.USER_INTERACTED);
+}
